Move views/search to /search
