+++ app/search/__init__.py
@@ -1,0 +1,2 @@
+from app.search.views import searchbp
+from app.search.forms import Search as SearchForm
