--- app/views/search/__init__.py
@@ -1,2 +1,0 @@
-from app.views.search.views import searchbp
-from app.views.search.forms import Search as SearchForm
